+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Added templating system for scaffolding Piral instances and pilets
 * Added Piral instance templates `@smapiot/piral-template-default` and `@smapiot/piral-template-empty`
 * Added pilet templates `@smapiot/pilet-template-default` and `@smapiot/pilet-template-empty`
+* Added `survey` option to Piral CLI commands
 
 ## 0.12.3 (October 10, 2020)